Eloquent Relationships
Using Eloquent Relationships we can relate two tables.

First we need to define the relationship between the two tables users
and jokes and then we can join the tables data.

So, our updated show method code will look like this, here first we are
joining User table data with the jokes tables, also we are finding the
previous and next joke and then we are adding them to the response.